feat: update app add box compo and update index css and update CarouselProducts
